# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Api
 * CloudVault personal cloud storage API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export type FileFileType = typeof FileFileType[keyof typeof FileFileType];


export const FileFileType = {
  image: 'image',
  video: 'video',
  document: 'document',
  pdf: 'pdf',
  other: 'other',
} as const;

export interface File {
  id: number;
  userId: string;
  name: string;
  objectPath: string;
  size: number;
  mimeType: string;
  fileType: FileFileType;
  /** @nullable */
  folderId?: number | null;
  isStarred: boolean;
  createdAt: string;
  updatedAt: string;
}

export type FileInputFileType = typeof FileInputFileType[keyof typeof FileInputFileType];


export const FileInputFileType = {
  image: 'image',
  video: 'video',
  document: 'document',
  pdf: 'pdf',
  other: 'other',
} as const;

export interface FileInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  objectPath: string;
  /** @minimum 0 */
  size: number;
  mimeType: string;
  fileType: FileInputFileType;
  /** @nullable */
  folderId?: number | null;
}

export interface FileUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  folderId?: number | null;
}

export interface Folder {
  id: number;
  userId: string;
  name: string;
  fileCount?: number;
  createdAt: string;
  updatedAt: string;
}

export interface FolderInput {
  /** @minLength 1 */
  name: string;
}

export interface FolderUpdate {
  /** @minLength 1 */
  name?: string;
}

export interface Note {
  id: number;
  userId: string;
  title: string;
  content: string;
  isPinned: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface NoteInput {
  /** @minLength 1 */
  title: string;
  content: string;
}

export interface NoteUpdate {
  /** @minLength 1 */
  title?: string;
  content?: string;
}

export type DashboardSummaryFilesByType = {
  image: number;
  video: number;
  document: number;
  pdf: number;
  other: number;
};

export interface DashboardSummary {
  totalFiles: number;
  totalNotes: number;
  totalFolders: number;
  storageUsedBytes: number;
  filesByType: DashboardSummaryFilesByType;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  file: 'file',
  note: 'note',
} as const;

export type ActivityItemAction = typeof ActivityItemAction[keyof typeof ActivityItemAction];


export const ActivityItemAction = {
  created: 'created',
  updated: 'updated',
  deleted: 'deleted',
  starred: 'starred',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  entityType: string;
  name: string;
  action: ActivityItemAction;
  createdAt: string;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

/**
 * Unauthorized
 */
export type UnauthorizedResponse = ErrorEnvelope;

/**
 * Bad request
 */
export type BadRequestResponse = ErrorEnvelope;

/**
 * Not found
 */
export type NotFoundResponse = ErrorEnvelope;

export type ListFilesParams = {
/**
 * @nullable
 */
folderId?: number | null;
search?: string;
type?: ListFilesType;
};

export type ListFilesType = typeof ListFilesType[keyof typeof ListFilesType];


export const ListFilesType = {
  image: 'image',
  video: 'video',
  document: 'document',
  pdf: 'pdf',
  other: 'other',
} as const;

export type ListNotesParams = {
search?: string;
};