# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CloudVault personal cloud storage API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List files for the authenticated user
      parameters:
        - name: folderId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [image, video, document, pdf, other]
      responses:
        "200":
          description: List of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/File"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createFile
      tags: [files]
      summary: Register an uploaded file in the database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileInput"
      responses:
        "201":
          description: File registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/File"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /files/{id}:
    get:
      operationId: getFile
      tags: [files]
      summary: Get a single file by ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: File found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/File"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      operationId: updateFile
      tags: [files]
      summary: Rename or move a file
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileUpdate"
      responses:
        "200":
          description: File updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/File"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /files/{id}/star:
    patch:
      operationId: toggleFileStar
      tags: [files]
      summary: Toggle starred status of a file
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Star toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/File"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Folders ────────────────────────────────────────────────────────────────
  /folders:
    get:
      operationId: listFolders
      tags: [folders]
      summary: List all folders for the authenticated user
      responses:
        "200":
          description: List of folders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Folder"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createFolder
      tags: [folders]
      summary: Create a new folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FolderInput"
      responses:
        "201":
          description: Folder created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /folders/{id}:
    patch:
      operationId: updateFolder
      tags: [folders]
      summary: Rename a folder
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FolderUpdate"
      responses:
        "200":
          description: Folder updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      operationId: deleteFolder
      tags: [folders]
      summary: Delete a folder (files inside are moved to root)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Notes ──────────────────────────────────────────────────────────────────
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List all notes for the authenticated user
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a new note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Note created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /notes/{id}:
    get:
      operationId: getNote
      tags: [notes]
      summary: Get a single note
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Note found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdate"
      responses:
        "200":
          description: Note updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /notes/{id}/pin:
    patch:
      operationId: toggleNotePin
      tags: [notes]
      summary: Toggle pinned status of a note
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Pin toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Dashboard ──────────────────────────────────────────────────────────────
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── File ──────────────────────────────────────────────────────────────────
    File:
      type: object
      required: [id, userId, name, objectPath, size, mimeType, fileType, isStarred, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        objectPath:
          type: string
        size:
          type: integer
        mimeType:
          type: string
        fileType:
          type: string
          enum: [image, video, document, pdf, other]
        folderId:
          type: ["integer", "null"]
        isStarred:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    FileInput:
      type: object
      required: [name, objectPath, size, mimeType, fileType]
      properties:
        name:
          type: string
          minLength: 1
        objectPath:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 0
        mimeType:
          type: string
        fileType:
          type: string
          enum: [image, video, document, pdf, other]
        folderId:
          type: ["integer", "null"]
    FileUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        folderId:
          type: ["integer", "null"]

    # ── Folder ────────────────────────────────────────────────────────────────
    Folder:
      type: object
      required: [id, userId, name, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        fileCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    FolderInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
    FolderUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1

    # ── Note ──────────────────────────────────────────────────────────────────
    Note:
      type: object
      required: [id, userId, title, content, isPinned, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        content:
          type: string
        isPinned:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    NoteInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
    NoteUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string

    # ── Dashboard ─────────────────────────────────────────────────────────────
...