# lib/api-zod/src/generated/types/errorEnvelope.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Api
 * CloudVault personal cloud storage API
 * OpenAPI spec version: 0.1.0
 */

export interface ErrorEnvelope {
  error: string;
}